+++ packages/deck/src/datasets/preparedDatasetStore.ts
@@ -1,45 +1,23 @@
-import type * as arrow from 'apache-arrow';
-import type {QueryHandle} from '@sqlrooms/duckdb';
 import {createStore} from 'zustand/vanilla';
+import type {QueryHandle} from '@sqlrooms/duckdb';
 import {
   isSqlDatasetInput,
   type DeckDatasetInput,
   type PreparedDeckDatasetState,
 } from '../types';
 import {prepareDeckDataset} from '../prepare/prepareDeckDataset';
-import type {GeometryEncodingHint, PreparedDeckDataset} from '../prepare/types';
 import {resolveArrowTable} from './normalizeDatasets';
-
-const DEFAULT_MAX_PREPARED_DATASET_ENTRIES = 20;
-
-let nextDatasetAccess = 0;
-let nextTableIdentity = 0;
-let nextSqlSourceIdentity = 0;
-
-const tableIdentities = new WeakMap<arrow.Table, string>();
-const sqlSourceIdentities = new WeakMap<object, string>();
-
-/**
- * One cached prepared-dataset entry keyed by resolved dataset identity.
- *
- * The entry tracks both lifecycle state and eviction metadata:
- *
- * - `consumers` records which hook instances currently reference the entry, so
- *   live entries are not evicted while a map is still using them
- * - `lastAccessedAt` drives the internal LRU policy for settled entries
- * - the discriminated union stores either the in-flight preparation promise,
- *   the prepared dataset, or the terminal error
- */
-type PreparedDatasetCacheEntry = {
-  /** Consumer ids currently subscribed to this cache entry. */
-  consumers: Set<string>;
-  /** Monotonic access counter used to evict the least-recently-used entry. */
-  lastAccessedAt: number;
-} & (
-  | {status: 'loading'; promise: Promise<void>}
-  | {status: 'ready'; prepared: PreparedDeckDataset}
-  | {status: 'error'; error: Error}
-);
+import {
+  DEFAULT_MAX_PREPARED_DATASET_ENTRIES,
+  cloneEntryWithConsumers,
+  evictLruEntries,
+  nextAccessTimestamp,
+  touchEntry,
+} from './helpers';
+import type {
+  PreparedDatasetCacheEntry,
+  PreparedDatasetStoreOptions,
+} from './types';
 
 /**
  * Internal store state for the prepared-dataset cache.
@@ -97,195 +75,6 @@ type PreparedDatasetStoreState = {
    */
   syncConsumer: (consumerId: string, keys: string[]) => void;
 };
-
-type PreparedDatasetStoreOptions = {
-  /** Maximum number of settled prepared entries to retain before LRU eviction. */
-  maxEntries?: number;
-  /** Injectable preparation function for tests and future specialization. */
-  prepareDataset?: (options: {
-    datasetId: string;
-    geometryColumn?: string;
-    geometryEncodingHint?: GeometryEncodingHint;
-    table: arrow.Table;
-  }) => PreparedDeckDataset;
-};
-
-function nextAccessTimestamp(): number {
-  nextDatasetAccess += 1;
-  return nextDatasetAccess;
-}
-
-/**
- * Assign a stable cache identity to an Arrow table object.
- *
- * Table identities are stored in a `WeakMap` so this cache does not extend the
- * lifetime of tables once they are no longer referenced elsewhere.
- */
-function getTableIdentity(table: arrow.Table): string {
-  const cached = tableIdentities.get(table);
-  if (cached) {
-    return cached;
-  }
-
-  nextTableIdentity += 1;
-  const identity = `table:${nextTableIdentity}`;
-  tableIdentities.set(table, identity);
-  return identity;
-}
-
-/**
- * Assign a stable cache identity to the current SQL source object.
- *
- * For SQL datasets we include the upstream DuckDB connector identity in the
- * cache key so two rooms using the same SQL text do not accidentally share
- * prepared results across different database instances.
- */
-function getSqlSourceIdentity(sqlSourceIdentity: object): string {
-  const cached = sqlSourceIdentities.get(sqlSourceIdentity);
-  if (cached) {
-    return cached;
-  }
-
-  nextSqlSourceIdentity += 1;
-  const identity = `sql-source:${nextSqlSourceIdentity}`;
-  sqlSourceIdentities.set(sqlSourceIdentity, identity);
-  return identity;
-}
-
-function buildGeometryKey(input: DeckDatasetInput): string {
-  return `${input.geometryColumn ?? ''}\u0001${input.geometryEncodingHint ?? ''}`;
-}
-
-/**
- * Build the canonical cache key for one dataset input.
- *
- * The key intentionally ignores the user-facing dataset id and instead uses
- * the underlying data identity:
- *
- * - SQL datasets: DuckDB connector identity + SQL text + geometry options
- * - Arrow datasets: table object identity + geometry options
- *
- * Unresolved Arrow inputs (`arrowTable: undefined`) return `undefined`, which
- * signals that the dataset should remain in `loading` until a table exists.
- */
-export function resolvePreparedDatasetCacheKey(options: {
-  input: DeckDatasetInput;
-  sqlSourceIdentity?: object;
-}): string | undefined {
-  const {input, sqlSourceIdentity} = options;
-
-  if (isSqlDatasetInput(input)) {
-    if (!sqlSourceIdentity) {
-      throw new Error(
-        'SQL dataset cache keys require a sqlSourceIdentity object.',
-      );
-    }
-
-    return [
-      'sql',
-      getSqlSourceIdentity(sqlSourceIdentity),
-      input.sqlQuery,
-      buildGeometryKey(input),
-    ].join('\u0001');
-  }
-
-  const table = resolveArrowTable(input);
-  if (!table) {
-    return undefined;
-  }
-
-  return ['arrow', getTableIdentity(table), buildGeometryKey(input)].join(
-    '\u0001',
-  );
-}
-
-/**
- * Clone a cache entry while replacing its consumer set and refreshing the
- * access timestamp.
- *
- * This keeps the discriminated union shape intact while centralizing the
- * bookkeeping needed by consumer reconciliation and LRU tracking.
- */
-function cloneEntryWithConsumers(
-  entry: PreparedDatasetCacheEntry,
-  consumers: Set<string>,
-): PreparedDatasetCacheEntry {
-  const lastAccessedAt = nextAccessTimestamp();
-
-  if (entry.status === 'loading') {
-    return {
-      status: 'loading',
-      promise: entry.promise,
-      consumers,
-      lastAccessedAt,
-    };
-  }
-
-  if (entry.status === 'ready') {
-    return {
-      status: 'ready',
-      prepared: entry.prepared,
-      consumers,
-      lastAccessedAt,
-    };
-  }
-
-  return {
-    status: 'error',
-    error: entry.error,
-    consumers,
-    lastAccessedAt,
-  };
-}
-
-/** Mark an entry as recently used without changing its lifecycle payload. */
-function touchEntry(
-  entry: PreparedDatasetCacheEntry,
-): PreparedDatasetCacheEntry {
-  return cloneEntryWithConsumers(entry, new Set(entry.consumers));
-}
-
-/**
- * Evict least-recently-used settled entries when the cache exceeds capacity.
- *
- * Only `ready` and `error` entries participate in eviction. `loading` entries
- * are preserved so in-flight work is not discarded, and referenced entries are
- * preserved so active maps never lose their prepared data mid-render.
- */
-function evictLruEntries(
-  entries: Record<string, PreparedDatasetCacheEntry>,
-  maxEntries: number,
-): Record<string, PreparedDatasetCacheEntry> {
-  const settledEntries = Object.entries(entries).filter(
-    ([, entry]) => entry.status !== 'loading',
-  );
-
-  if (settledEntries.length <= maxEntries) {
-    return entries;
-  }
-
-  const evictable = settledEntries
-    .filter(([, entry]) => entry.consumers.size === 0)
-    .sort(([, left], [, right]) => left.lastAccessedAt - right.lastAccessedAt);
-
-  if (evictable.length === 0) {
-    return entries;
-  }
-
-  const nextEntries = {...entries};
-  let settledCount = settledEntries.length;
-
-  for (const [cacheKey] of evictable) {
-    if (settledCount <= maxEntries) {
-      break;
-    }
-
-    delete nextEntries[cacheKey];
-    settledCount -= 1;
-  }
-
-  return nextEntries;
-}
 
 /**
  * Create a feature-local cache store for prepared deck datasets.
